remove obsolete models

+++ src/meltano/meltano/transform/models/historical_prices/historical_prices.sql
@@ -242,7 +242,7 @@ from (
                 adjusted_close,
                 case
                     when lag(adjusted_close) over wnd > 0
-                        then coalesce(adjusted_close / lag(adjusted_close) over wnd - 1, 0)
+                        then coalesce(adjusted_close::numeric / (lag(adjusted_close) over wnd)::numeric - 1, 0)
                     end                 as relative_daily_gain,
                 close,
                 date::date,
@@ -264,7 +264,7 @@ from (
                 c                                                               as adjusted_close,
                 case
                     when lag(c) over wnd > 0
-                        then coalesce(c / lag(c) over wnd - 1, 0)
+                        then coalesce(c::numeric / (lag(c) over wnd)::numeric - 1, 0)
                     end                                                         as relative_daily_gain,
                 c                                                               as close,
                 to_timestamp(t / 1000)::date                                    as date,
@@ -287,7 +287,7 @@ from (
                 c                                                                      as adjusted_close,
                 case
                     when lag(c) over wnd > 0
-                        then coalesce(c / lag(c) over wnd - 1, 0)
+                        then coalesce(c::numeric / (lag(c) over wnd)::numeric - 1, 0)
                     end                                                                as relative_daily_gain,
                 c                                                                      as close,
                 to_timestamp(t / 1000)::date                                           as date,
